use src as root folder, refactor directory names, add Dockerfile and requirements.txt
