+++ requirements.txt
@@ -1,0 +1,30 @@
+name: borrelai
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - apscheduler=3.5.3
+  - ca-certificates=2018.12.5
+  - certifi=2018.11.29
+  - click=7.0
+  - flask=1.0.2
+  - flask-login=0.4.1
+  - flask-sqlalchemy=2.3.2
+  - itsdangerous=1.1.0
+  - jinja2=2.10
+  - markupsafe=1.1.0
+  - openssl=1.1.1
+  - pip=18.1_0
+  - python=3.7.2
+  - pytz=2018.9
+  - setuptools=40.6.3
+  - six=1.12.0
+  - sqlalchemy=1.2.16
+  - sqlite=3.26.0
+  - tzlocal=1.5.1
+  - werkzeug=0.14.1
+  - wheel=0.32.3
+  - pip3:
+    - tinytag
+
+
