+++ src/flaskapp/app.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from flask import Flask, request, render_template, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import and_
-from db.objects import Songs, Votes, Round, SelectedSongs, IPAddress
+from ..db.objects import Songs, Votes, Round, SelectedSongs, IPAddress
 
 app = Flask(__name__)
 app.config.from_pyfile('.secrets')
