+++ src/flaskapp/dashboard/assets/_bootswatch.scss
