+++ src/flaskapp/dashboard/assets/_variables.scss
