+++ src/flaskapp/dashboard/assets/main.css
