+++ src/flaskapp/static/assets/particles.json
