+++ src/flaskapp/static/css/_variables.scss
