+++ src/flaskapp/static/img/logo_white.png
