+++ src/flaskapp/static/js/app.js
