+++ src/flaskapp/templates/vote.html
